[Improvement-14295][common] Remove the duplicate FileUtils
